fix:  rewards release period,  booster helper for boosted rate, other tax helper solidity library, fix calculations for rewards, claimed, staked, fix rounding issues and second-precision calculus, set token supply distribution on deploy, more params on deployment (taper threshold, all durations in seconds, fees in wei), added DonationEvent and WithdrawalEvent to calculate donations for any period in the past, fix type conversion (uint8-32-256), added isClaimable, taxTier and nextTaxTier functions, added stakedReward and unclaimedRewardBalance to LibGem.Gem struct, solidity-docgen, hh tasks update to test all that

+++ contracts/facets/GemGettersFacet.sol
@@ -42,7 +42,7 @@ contract GemGettersFacet {
 
         return block.timestamp - gemType.LastMint;
     }
-
+    ///TODO return false for unexisting gems
     function isMintAvailableForGem(uint8 _gemType) external view returns (bool) {
         LibGem.DiamondStorage storage ds = LibGem.diamondStorage();
         LibGem.GemTypeMetadata memory gemType = ds.GetGemTypeMetadata[_gemType];
@@ -56,7 +56,7 @@ contract GemGettersFacet {
         LibGem.Booster _booster,
         uint8 _type,
         address _user
-    ) public view returns (uint) {
+    ) public view returns (uint256) {
         LibUser.DiamondStorage storage ds = LibUser.diamondStorage();
         LibUser.UserData storage user = ds.GetUserData[_user];
         if (_booster == LibGem.Booster.Omega) {
